Unify cairo_run_test.rs
